Python Implementation of MCS using the GBM model
Basic application of Monte Carlo Simulation in Python. The model that
was used to specify the behavior of the stock price is the Geometric
Brownian Motion (GBM).

At the end of the simulation, 10,000 runs of random trials produced a
distribution of outcomes that can be analyzed.

The dataset used is a real-world dataset comprising of several stocks
from the Philippine Stocks Exchange (PSE) to illustrate the model.